Build fixes for ppc64 and x86_64 now that we use -Wshorten-64-to-32.
Canonical link: https://commits.webkit.org/13536@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@16156 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/API/testapi.c
@@ -846,8 +846,8 @@ static char* createStringWithContentsOfFile(const char* fileName)
 {
     char* buffer;
     
-    int buffer_size = 0;
-    int buffer_capacity = 1024;
+    size_t buffer_size = 0;
+    size_t buffer_capacity = 1024;
     buffer = (char*)malloc(buffer_capacity);
     
     FILE* f = fopen(fileName, "r");
